Move `HasLogicalModel` under `LogicalModel`
